Factor out initial piece placement
The HATETRIS algorithm should not have control over exactly where the
piece gets placed - it should just pick an index from the supplied
range. So, `placeFirstPiece` is now factored out to a place where the
algorithm can use it for its calculations but also the main game loop
can use it for actual piece placement.